OOoPy, Transformer
On deletion of an OOoPy object close it explicitly (uses __del__)
Ensure mimetype is the first element in the resulting archive, seems OOo
is picky about this.
When modifying the manifest make sure the manifest namespace is named
"manifest" not something auto-generated by ElementTree. I consider this
a bug in OOo to require this. This now uses the _namespace_map of
ElementTree and uses the same names as OOo for all namespaces.
When modifying (or creating) archive members, we create the OOo archive
as if it was a DOS system (type fat) and ensure we use the current
date/time (UTC). This also fixes problems with file permissions on newer
versions of pythons ZipFile.