Fix application of multiple Set_Attribute
If there are multiple Set_Attribute in a single Attribute_Access and
they transitively match an item we would apply multiple transformations
to the same tag/attribute combination, e.g., we would set for a
draw:line tag the text-style-name attribute first to "P1" and then to
"Concat_P1" if there are two Set_Attribute for the same tag/attr
combination with P13->P1 and P1->Concat_P1. This would even change with
the dictionary order... Now we apply only the first match.